# src/main/java/com/sunil/myretail/model/Product.java
package com.sunil.myretail.model;


import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Data;

@Data
@JsonInclude(JsonInclude.Include.NON_NULL)
@JsonIgnoreProperties(ignoreUnknown = true)
public class Product {

    @JsonProperty("id")
    private String productId;

    @JsonProperty("name")
    private String productDescription;

    @JsonProperty("current_price")
    private Price price;

}